manual config refactored into multiFile
