--- .gitignore
@@ -1,5 +1,0 @@
-monitorFunctions/__pycache__/monitorEpisodes.cpython-38.pyc
-monitorFunctions/__pycache__/monitorMaster.cpython-38.pyc
-monitorFunctions/__pycache__/monitorSeason.cpython-38.pyc
-__pycache__/config.cpython-38.pyc
-personal_config.py
